song_backend/flask_server/requirements.txt
Flask
transformers
torch
Pillow
requests